refactor: DataInitializer를 operation/utils 패키지로 이동
- 패키지 이동: app.bottlenote → app.bottlenote.operation.utils
- 다른 테스트 유틸리티(TestAuthenticationSupport, TestContainersConfig)와 동일한 위치로 통합
- IntegrationTestSupport에서 import 추가
