+++ bottlenote-product-api/src/test/java/app/bottlenote/operation/utils/DataInitializer.java
@@ -1,4 +1,4 @@
-package app.bottlenote;
+package app.bottlenote.operation.utils;
 
 import static jakarta.transaction.Transactional.TxType.REQUIRES_NEW;
